requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
requests==2.25.1
urllib3==1.26.3
Werkzeug==1.0.1
flask_sqlalchemy
wtforms
wtforms.validators
Flask-MySQLdb
pycryptodome==3.10.1
pytest==6.2.2
python-dotenv==0.15.0
Flask-Migrate==2.7.0
flask-script==2.0.6
black==20.8b1